# gateway/cmd/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Volha gateway REST API",
        "contact": {
            "name": "Alex \"bustard\" Provor",
            "url": "https://breezy.su",
            "email": "help@breezy.su"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/country/create": {
            "post": {
                "description": "Добавляет новую страну",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "country"
                ],
                "summary": "Создать страну",
                "parameters": [
                    {
                        "description": "Новая страна",
                        "name": "country",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.Country"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Страна успешно создана",
                        "schema": {
                            "$ref": "#/definitions/views.SWGSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/country/delete": {
            "delete": {
                "description": "Удаляет страну по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "country"
                ],
                "summary": "Удалить страну",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID страны",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Страна успешно удалена",
                        "schema": {
                            "$ref": "#/definitions/views.SWGSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/country/getall": {
            "get": {
                "description": "Возвращает список всех стран",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "country"
                ],
                "summary": "Получить все страны",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "schema": {
                            "$ref": "#/definitions/views.SWGCountryListResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/country/update": {
            "put": {
                "description": "Обновляет данные существующей страны",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "country"
                ],
                "summary": "Обновить страну",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID страны",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Обновлённые данные страны",
                        "name": "country",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.Country"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Страна успешно обновлена",
                        "schema": {
                            "$ref": "#/definitions/views.SWGSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID или данные",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/dictionaries/getall": {
            "get": {
                "description": "Возвращает список всех доступных справочных данных: бренды, категории, материалы, страны и цвета",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dictionaries"
                ],
                "summary": "Получить все справочники",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "schema": {
                            "$ref": "#/definitions/views.Dictionaries"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/dictionaries/getall/category": {
            "get": {
                "description": "Возвращает список всех доступных справочных данных: бренды, материалы, страны и цвета и размеры по определенной категории",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dictionaries"
                ],
                "summary": "Получить все справочники по категории",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID категории",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "schema": {
                            "$ref": "#/definitions/views.Dictionaries"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "views.Brand": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "views.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
        "views.Color": {
            "type": "object",
            "properties": {
                "hex": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "views.Country": {
            "type": "object",
            "properties": {
                "friendly": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "views.Dictionaries": {
            "type": "object",
            "properties": {
                "brands": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/views.Brand"
                    }
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/views.Category"
                    }
                },
                "colors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/views.Color"
                    }
                },
                "countries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/views.Country"
                    }
                },
                "materials": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/views.Material"
                    }
                },
                "max_depth": {
                    "type": "integer"
                },
                "max_height": {
                    "type": "integer"
                },
                "max_price": {
                    "type": "integer"
                },
                "max_width": {
                    "type": "integer"
                },
                "min_depth": {
                    "type": "integer"
                },
                "min_height": {
                    "type": "integer"
                },
                "min_price": {
                    "type": "integer"
                },
                "min_width": {
                    "type": "integer"
                }
            }
        },
        "views.Material": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "views.SWGCountryListResponse": {
            "type": "object",
            "properties": {
                "countries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/views.Country"
                    }
                }
            }
        },
        "views.SWGErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "something went wrong"
                }
            }
        },
...
        "views.SWGSuccessResponse": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string",
                    "example": "operation successful"
                }
            }
        }
    }
}